Init new 'test' branch push.
Testing the abilities of GitHub's version control by reversing the
colors on 'test.html' within the repo. Need to be able to see this
version on this branch, then 'git checkout master' and refresh the page
locally once more to see the colors switch back. If it works, GitHub is
functional within the workflow!!